Refactored Search to render children and SearchBar
