+++ src/containers/SearchContainer/SearchContainer.jsx
@@ -1,13 +1,16 @@
-import SearchResultsContainer from './SearchResultsContainer';
-import SearchBar from 'components/search';
+import SearchResultsContainer from './SearchResults';
+import DropdownContainer from './Dropdown';
+import Search from 'components/search';
 
-const SearchContainer = ({ fetchGeoDataFromZip, searchResult, currentUser }) => {
-    return (
-        <div className="search-container">
-            <SearchBar fetchGeoDataFromZip={fetchGeoDataFromZip} />
-            <SearchResultsContainer currentUser={currentUser} searchResult={searchResult} />
-        </div>
-    );
+const SearchContainer = ({ fetchGeoDataFromZip, searchResult, currentUser, setCurrentUser }) => {
+  return (
+      <div className="search-container">
+          <Search fetchGeoDataFromZip={fetchGeoDataFromZip}>
+            <DropdownContainer currentUser={currentUser} setCurrentUser={setCurrentUser} />
+          </Search>
+          <SearchResultsContainer currentUser={currentUser} searchResult={searchResult} />
+      </div>
+  );
 }
 
 export default SearchContainer;
