+++ src/containers/SearchContainer/SearchResults/index.js
@@ -1,0 +1,3 @@
+import SearchResultsContainer from './SearchResultsContainer';
+
+export default SearchResultsContainer;
